requirements.txt
fastapi>=0.95.0,<0.96
uvicorn>=0.21.1,<0.22
tortoise-orm>=0.19.3,<0.20
pydantic[email]>=1.10.7,<1.11
python-jose[cryptography]>=3.3.0,<3.4
passlib[bcrypt]>=1.7.4,<1.8
python-multipart>=0.0.5,<0.1
